Simplify and improve axes latex tests
