+++ bokehjs/test/integration/_util.ts
@@ -1,11 +1,15 @@
+import sinon from "sinon"
+
 export {describe, it, display} from "../framework"
 
 import {Matrix} from "@bokehjs/core/util/matrix"
 import {Figure, figure} from "@bokehjs/api/plotting"
 import {LayoutDOM, Row, Column, GridBox} from "@bokehjs/models/layouts/index"
+
 import {wait} from "@bokehjs/core/util/defer"
 import {tex2svg, mathml2svg} from "@bokehjs/models/text/mathjax"
 import {MathJaxProvider} from "@bokehjs/models/text/providers"
+import {MathTextView} from "@bokehjs/models/text/math_text"
 
 export function grid(items: Matrix<LayoutDOM> | LayoutDOM[][], opts?: Partial<GridBox.Attrs>): GridBox {
   const children = Matrix.from(items).to_sparse()
@@ -46,3 +50,17 @@ export class InternalProvider extends MathJaxProvider {
     this.status = "loaded"
   }
 }
+
+export function with_provider(provider: MathJaxProvider) {
+  return async (fn: () => Promise<void>) => {
+    const stub = sinon.stub(MathTextView.prototype, "provider")
+    stub.value(provider)
+    try {
+      await fn()
+    } finally {
+      stub.restore()
+    }
+  }
+}
+
+export const with_internal = with_provider(new InternalProvider())
